+++ src/design-system/Button/index.ts
@@ -1,2 +1,6 @@
-export { Button } from "./Button";
+export {
+  Button,
+  BUTTON_SIZE,
+  BUTTON_VARIANT,
+} from "./Button";
 export type { ButtonVariant, ButtonSize } from "./Button";